session dont save default menu labels and icons
This will make session files smaller and will allow updated translations of
menu item names to be seen so long as the user did not change the name.

NOTE:  As a result of this change, when downgrading to older spacefm versions,
customized menu item names and icons of built-in items will be lost (custom
menu items will not be affected).