revert desktop dialogs pop to top #424 bba93a50 dlgs on wrong desktop
bba93a50 set NULL parent window for desktop dialogs, but this change caused
some dialogs to appear on the wrong desktop (eg with openbox).  This commit
reverts those changes so desktop dialogs are again parented by the root
window.

WMs that refuse to pop dialogs to the top when parented by the root window
probably need to be fixed.