[GTK2] fix black background dir load regression 888a7735 #578 #627
Revised code adds back in a dropped gdk_window_set_background call
which had caused the widget background to be black prior to directory
contents loading.

Apparently setting the style directly on the widget causes a slowdown,
so keeping the old broken method that results in the ExoIconView not
restyling when the user changes a style. For more detail:

https://github.com/IgnorantGuru/spacefm/issues/627