Add support for Btrfs clone ioctl
Btrfs clone ioctl can be used for O(1) copying within a file system
utilizing copy-on-write.

Passing --enable-btrfs-clone to ./configure enables this feature.

The feature is dependent on the "btrfs/ioctl.h" header which is part of
btrfs-progs.

vfs_file_task_do_copy is modified to always try to call the ioctl,
and fall back to the regular copying method if it fails.