fix drag & drop segfaults #636 [BwackNinja]
Per BwackNinja: gtk_tree_drag_source_drag_data_delete should be wrapped in a
if (GTK_IS_TREE_DRAG_SOURCE (model)) because the drag source is not always a
tree, for example, dragging from the icon view to the tree sidebar. Other uses
of gtk_tree_drag_source_* are similarly wrapped in the same file.